Update jenkins-pipeline-docker-ansible-user-choice
This pipeline is designed to build and deploy two services (Service 1 and Service 2) by automating the process of creating Dockerfiles, building Docker images, pushing them to DockerHub, and deploying them using Ansible. The user can choose which service to build and deploy by selecting it from a list of choices. The pipeline uses environment variables to store sensitive information such as the DockerHub username and password, and it follows a consistent structure for building and deploying both services.